Refactor: usecaseの技術的要素の境界面の置き場所を変更
