+++ internal/controller/job/usercount/user_count_job.go
@@ -6,7 +6,7 @@ import (
 
 	"boilerplate-go/internal/logging"
 	"boilerplate-go/internal/observability"
-	"boilerplate-go/internal/usecase/interface/job"
+	"boilerplate-go/internal/usecase/boundary/job"
 	"boilerplate-go/internal/usecase/user"
 	"boilerplate-go/pkg/ptr"
 )
